Move FullStack fixure into full_space_spec.rb
Also adapt the routes to use `scope` instead of
repeated prefix "poll/"
